+++ client/src/app/components/maestros-list/maestros-list.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Chart } from 'chart.js';
 import { MaestrosService } from 'src/app/services/maestros/maestros.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-maestros-list',
@@ -32,10 +33,35 @@ export class MaestrosListComponent {
   }
 
   deleteMaestro(id_maestro: number){
-    this.maestrosService.deleteMaestro(id_maestro).subscribe(
-      (res) => console.log(res)
-      
-    )
-    this.getMaestro();
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.maestrosService.deleteMaestro(id_maestro).subscribe(
+          (res) => console.log(res),
+          err =>{Swal.fire({
+            icon: "error",
+            title: "Oops...",
+            text: "Something went wrong!",
+            footer: '<a href="#">Why do I have this issue?</a>'
+          });
+        } 
+          
+        )
+        this.getMaestro();
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success"
+        });
+      }
+    });
+ 
   }
 }
