+++ client/src/app/components/users/users.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { UsersService } from '../../services/users/users.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-users',
@@ -31,13 +32,33 @@ export class UsersComponent implements OnInit {
   }
 
   deleteUser(id_user : string){
-    this.usersService.deleteUser(id_user).subscribe(
-      (res) => {
-        console.log(res);
-        this.getUsers();
-      },
-      (err) => console.error(err)
-
-    );
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.usersService.deleteUser(id_user).subscribe(
+          (res) => {
+            console.log(res);
+            this.getUsers();
+          },
+          (err) => console.error(err)
+    
+        );
+        Swal.fire({
+          title: "Deleted!",
+          
+          text: "The user has been deleted.",
+          icon: "success"
+        });
+      }
+      
+    });
+   
   }
 }
